refactor: removing unnecessary unboxing

+++ src/main/java/module/lineup/substitution/WhenTextField.java
@@ -95,7 +95,7 @@ public class WhenTextField extends JFormattedTextField {
 		@Override
 		public String valueToString(Object obj) throws ParseException {
 			Integer value = (Integer) obj;
-			if (value == null || value.intValue() < 0) {
+			if (value == null || value < 0) {
 				return noValueDisplayString;
 			}
 			return MessageFormat.format(valueDisplayString, value);
